requirements.txt
fastapi
mangum
python-docx
pdfplumber
python-multipart
lxml
reportlab
uvicorn